Alternatives for specifying the font file name
If the drawtext filter is used more than once, a font file
specified in the first occurrence becomes the default that
subsequent invocations of the filter can use by omitting their
own fontfile specifications.

In addition, an environment variable, FFMPEG_DRAWTEXT_FONTFILE, is
used if it exists and no font file has yet been specified on the
command line. That file then becomes the default.

Signed-off-by: Ray Simard <rhs.ffmpeg@sylvan-glade.com>