Divide-by-zero vulnerability: stopgap
Option -qmin 0 causes a divide-by-zero exception here. I'm not sure what the
most appropriate solution is, so this patch is a temporary measure, simply an
assertion to give more concrete and useful information to the user than the
simple floating-point exception message (and is intentionally (qscale != 0)
rather than (qscale) for that reason).

Signed-off-by: Ray Simard <rhs.ffmpeg@sylvan-glade.com>