Zoom-related code removed; minor cleanups
The transform code included a zoom factor the purpose of which
was unclear.  Presumably it was meant as another way of handling
the vacated space at the edges, by enlarging the image to fill the
gaps. Unfortunately, doing the zoom in the transform and then
subsequently saving the result as the reference frame for the next
transform gives the filter fits as it tries to correct for one frame
what it did to the previous one.

Users who want this approach can chain scale and crop filters after
deshake. Perhaps, if the convenience seems worth it, some way to
make that automatic would be nice, or to add a second transform to
deshake after the de-shaken frame has been saved as the reference frame.

Signed-off-by: Ray Simard <rhs.ffmpeg@sylvan-glade.com>